Created GridConstructor class
I removed the grid construction logic from the Board class and added it
to a new GridConstructor class.

The Board object now stores a list of Betting space objects which will
be needed for the win/loss code.
